Featured images, elementor bug fix, translations
Added options to use the posts featured image for FB and TW by default.
If FB or TW image is added, it will use that instead of the featured image.
Fixed a bug that was preventing the media loader from working in Elementor.
More translation options with the use of __() instead of just echoing. More coming.